production_requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-coreapi==2.3.0
django-cors-headers==2.4.0
django-excel==0.0.10
django-extensions==2.1.5
django-request-logging==0.6.5
django-model-utils==3.2.0
djangorestframework==3.9.1
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
lazy-object-proxy==1.3.1
lml==0.0.9
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.17.0
odfpy==1.3.5
openpyxl==2.5.14
pandas==0.25.0
pycodestyle==2.5.0
pyexcel==0.5.10
pyexcel-io==0.5.11
pyexcel-ods==0.5.4
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
PyJWT==1.7.1
pylint==2.2.2
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
texttable==1.6.0
typed-ast==1.3.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
django-auth-ldap==1.7.0
python-ldap==3.1.0
